refactor: 메서드명 및 패키지명 변경
