+++ src/main/java/wooteco/subway/service/SectionService.java
@@ -1,7 +1,10 @@
 package wooteco.subway.service;
 
+import java.util.List;
+
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
 import wooteco.subway.dao.LineDao;
 import wooteco.subway.dao.SectionDao;
 import wooteco.subway.dao.StationDao;
@@ -9,7 +12,6 @@ import wooteco.subway.domain.Section;
 import wooteco.subway.domain.Sections;
 import wooteco.subway.domain.Station;
 import wooteco.subway.exception.DataNotFoundException;
-import java.util.List;
 
 @Service
 public class SectionService {
@@ -52,7 +54,7 @@ public class SectionService {
     }
 
     @Transactional(readOnly = true)
-    public List<Station> getStationsByLine(final long lineId) {
+    public List<Station> findStationsByLine(final long lineId) {
         final List<Section> lineSections = sectionDao.findAllByLineId(lineId);
         final Sections sections = new Sections(lineSections);
         return sections.extractStations();
