+++ src/main/java/wooteco/subway/service/StationService.java
@@ -1,11 +1,13 @@
 package wooteco.subway.service;
 
+import java.util.List;
+
 import org.springframework.stereotype.Service;
+
 import wooteco.subway.dao.StationDao;
 import wooteco.subway.domain.Station;
 import wooteco.subway.exception.DataNotFoundException;
 import wooteco.subway.exception.DuplicateNameException;
-import java.util.List;
 
 @Service
 public class StationService {
@@ -21,7 +23,7 @@ public class StationService {
         return stationDao.save(station);
     }
 
-    public List<Station> getAllStations() {
+    public List<Station> findAllStations() {
         return stationDao.findAll();
     }
